refactor: move cliente form and drop bcrypt
