+++ forms.py
@@ -1,5 +1,12 @@
 from flask_wtf import FlaskForm, RecaptchaField
-from wtforms import StringField, FloatField, IntegerField, SelectMultipleField, PasswordField
+from wtforms import (
+    StringField,
+    FloatField,
+    IntegerField,
+    SelectMultipleField,
+    PasswordField,
+    SubmitField,
+)
 from wtforms.validators import DataRequired, Optional, Email
 
 class TipoInscricaoEventoForm(FlaskForm):
@@ -10,6 +17,13 @@ class RegraInscricaoEventoForm(FlaskForm):
     tipo_inscricao_id = IntegerField('Tipo de Inscrição', validators=[DataRequired()])
     limite_oficinas = IntegerField('Limite de Oficinas', validators=[Optional()])
     oficinas_permitidas = SelectMultipleField('Oficinas Permitidas', coerce=int)
+
+
+class EditarClienteForm(FlaskForm):
+    nome = StringField('Nome', validators=[DataRequired()])
+    email = StringField('E-mail', validators=[DataRequired(), Email()])
+    senha = PasswordField('Nova Senha')
+    submit = SubmitField('Salvar Alterações')
 
 
 class PublicClienteForm(FlaskForm):
