+++ requirements
@@ -1,6 +1,5 @@
 alembic==1.14.1
 APScheduler==3.11.0
-bcrypt==4.3.0
 bidict==0.23.1
 bleach==6.2.0
 blinker==1.9.0
